requirements.txt
streamlit==1.30.0
tensorflow==2.14.0
pillow==10.0.0
numpy==1.27.0
matplotlib==3.8.0
scikit-learn==1.3.0
seaborn==1.6.0
pandas==2.2.0